@@ -1,59 +1,47 @@
 {
   "name": "openchat",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "generate-chats-qr": "node src/generate_chats_qr.js"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "lint-staged",
+    "lint:es": "eslint \"*.{js,vue}\" \"src/**/*.{js,vue}\"",
+    "lint:es:fix": "eslint \"*.{js,vue}\" \"src/**/*.{js,vue}\" --fix",
+    "prepare": "husky install"
+  },
+  "lint-staged": {
+    "*.{js,vue}": "eslint"
   },
   "dependencies": {
-    "autoprefixer": "^10",
-    "core-js": "^3.8.3",
-    "dayjs": "^1.11.12",
+    "@heroicons/vue": "^2.0.18",
+    "dayjs": "^1.11.13",
     "jwt-decode": "^4.0.0",
-    "nanoid": "^5.0.7",
-    "postcss": "^8",
-    "tailwindcss": "^3",
-    "ua-parser-js": "^1.0.38",
-    "vue": "^2.7.16",
-    "vue-router": "^3.5.1"
+    "ky": "^1.7.2",
+    "ua-parser-js": "^1.0.39",
+    "vue": "^3.5.12",
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@tailwindcss/typography": "^0.5.2",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "axios": "^1.7.7",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "images": "^3.2.4",
-    "qrcode": "^1.5.0",
-    "vue-cli-plugin-axios-x": "~0.2.6",
-    "vue-cli-plugin-tailwind": "~3.0.0",
-    "vue-template-compiler": "^2.7.16"
+    "@commitlint/cli": "^17.4.4",
+    "@commitlint/config-conventional": "^17.4.4",
+    "@tailwindcss/typography": "^0.5.15",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "autoprefixer": "^10.4.15",
+    "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^9.12.0",
+    "eslint-plugin-vue": "^9.29.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^15.2.10",
+    "postcss": "^8.4.29",
+    "tailwindcss": "^3.3.3",
+    "vite": "^4.4.5"
   },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "@babel/eslint-parser"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
+  }
 }
