@@ -1,13 +1,10 @@
-import Vue from 'vue'
-import './plugins/axios'
-import './plugins/profile'
-import App from './App.vue'
-import router from './router'
-import './assets/tailwind.css'
+import { createApp } from "vue"
+import App from "./App.vue"
 
-Vue.config.productionTip = false
+import router from "./router";
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+import "./style.css"
+
+createApp(App).
+    use(router).
+    mount("#app");
