@@ -1,16 +1,12 @@
 <template>
   <div class="relative bg-white">
-    <app-header/>
-    <router-view/>
-    <app-footer/>
+    <app-header />
+    <router-view />
+    <app-footer />
   </div>
 </template>
 
-<script>
-import AppHeader from "@/components/AppHeader"
-import AppFooter from "@/components/AppFooter"
-export default {
-  name: 'App',
-  components: {AppHeader, AppFooter},
-}
+<script setup>
+import AppHeader from "./components/AppHeader.vue"
+import AppFooter from "./components/AppFooter.vue"
 </script>
