@@ -1,25 +1,19 @@
 <template>
-  <div class="mt-10 mx-auto py-10 max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
+  <div class="mt-10 mx-auto py-10 max-w-7xl px-12">
     <div class="sm:text-center lg:text-left">
-      <h1 class="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl">
-        <span class="block text-sky-600 xl:inline">404</span>
+      <h1 class="text-4xl tracking-tight font-bold text-gray-900">
+        <span class="block text-sky-600 xl:inline md:mr-3">404</span>
         <span class="block xl:inline">頁面不存在</span>
       </h1>
       <p class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0">
         您所請求的頁面不見了，或者是從來沒有存在過。
       </p>
-      <div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
-        <div class="rounded-md shadow">
-          <button v-if="$router.history.length"
-                  class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white bg-sky-600 hover:bg-sky-700 md:py-4 md:text-lg md:px-10"
-                  @click="$router.back">
-            返回上一頁
-          </button>
-        </div>
+      <div class="mt-5 sm:mt-16 sm:flex sm:justify-center lg:justify-start">
         <div class="mt-3 sm:mt-0 sm:ml-3">
           <router-link
-              class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-sky-700 bg-sky-100 hover:bg-sky-200 md:py-4 md:text-lg md:px-10"
-              to="/">
+            class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-sky-700 bg-sky-100 hover:bg-sky-200 md:py-4 md:text-lg md:px-10"
+            to="/"
+          >
             回首頁
           </router-link>
         </div>
